Refactor TestLocalExecutionPlanner for better reuse of mocked classes Moved OutputBuffer class and RemoteSourceFactory to public scope
