+++ presto-main/src/test/java/com/facebook/presto/sql/planner/TestingRemoteSourceFactory.java
@@ -1,0 +1,62 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.sql.planner;
+
+import com.facebook.presto.Session;
+import com.facebook.presto.common.block.SortOrder;
+import com.facebook.presto.common.type.Type;
+import com.facebook.presto.operator.DriverContext;
+import com.facebook.presto.operator.SourceOperator;
+import com.facebook.presto.operator.SourceOperatorFactory;
+import com.facebook.presto.spi.plan.PlanNodeId;
+
+import java.util.List;
+
+public class TestingRemoteSourceFactory
+        implements RemoteSourceFactory
+{
+    @Override
+    public SourceOperatorFactory createRemoteSource(Session session, int operatorId, PlanNodeId planNodeId, List<Type> types)
+    {
+        return new TestingSourceOperatorFactory();
+    }
+
+    @Override
+    public SourceOperatorFactory createMergeRemoteSource(Session session, int operatorId, PlanNodeId planNodeId, List<Type> types, List<Integer> outputChannels, List<Integer> sortChannels, List<SortOrder> sortOrder)
+    {
+        throw new UnsupportedOperationException();
+    }
+
+    private static class TestingSourceOperatorFactory
+            implements SourceOperatorFactory
+    {
+        @Override
+        public PlanNodeId getSourceId()
+        {
+            return new PlanNodeId("test");
+        }
+
+        @Override
+        public SourceOperator createOperator(DriverContext driverContext)
+        {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public void noMoreOperators()
+        {
+            throw new UnsupportedOperationException();
+        }
+    }
+}
